Organize types into separate files by category
- Split src/types/index.ts into 5 logical files:
  - plan.ts: Health plan related types (HealthPlan, PlanPremiums, etc.)
  - user.ts: User input related types (UserInputs, UserCosts, etc.)
  - results.ts: Calculation result types (PlanResult)
  - url.ts: URL parameter types (URLParams)
  - data.ts: Plan data structure types (PlanData, contribution limits)
- Updated index.ts to re-export all types maintaining backward compatibility
- Verified build and tests still pass with new structure
