+++ src/types/url.ts
@@ -1,0 +1,8 @@
+export interface URLParams {
+  year?: string;
+  coverage?: string;
+  ageGroup?: string;
+  taxRate?: string;
+  hsaContribution?: string;
+  fsaContribution?: string;
+}
